pyproject.toml
[tool.poetry]
name = "blue_horizon"
version = "0.1.0"
description = "Blue Horizon AI Concierge - An agentic AI assistant for a luxury hotel."
authors = ["Emmanuel Awa <emmanuel@emmanuelawa.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^1.4.0"
numpy = "^1.26.0"
faker = "^8.0.0"
openai = "^1.58.1"
tqdm = "^4.65.0"
tenacity = "^8.2.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.4"
inflection = "^0.5.1"
llama-index = "^0.12.10"
chromadb = "^0.6.1"
spacy = "^3.8.3"
python-dateutil = "^2.9.0.post0"
redis = "^5.0.0"
llama-index-llms-anthropic = "^0.6.3"
llama-index-llms-gemini = "^0.4.2"
google-cloud-aiplatform = "^1.75.0"
google-generativeai = "^0.8.3"
pytest-cov = "^6.0.0"
langchain = "^0.3.13"
pylint = "^3.3.3"
requests = "^2.32.3"
timezonefinder = "^6.5.7"
geopy = "^2.4.1"
pytz = "^2024.2"
certifi = "^2024.12.14"
ag2 = "^0.6.1"
sentence-transformers = "^2.5.1"
torch = "2.2.2"
onnxruntime = "1.19.2" 
llama-index-embeddings-openai = "^0.3.1"
llama-index-vector-stores-chroma = "^0.4.1"
streamlit = "^1.41.1"
flaml = {extras = ["automl"], version = "^2.3.3"}
nest-asyncio = "^1.6.0"
tiktoken = "0.6.0"
asyncpg = "^0.30.0"
python-magic = "^0.4.27"
tabulate = "^0.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.scripts]
concierge = "blue_horizon.cli:main"
generate-reviews = "blue_horizon.data.generate_synthetic_data_llm:generate_synthetic_reviews"
GenerateGretelSyntheticData = "blue_horizon.data.generate_synthetic_data_gretel:main"
GenerateSyntheticDatasets = "blue_horizon.data.generate_synthetic_data:main"
AugmentSyntheticDataWithGretel = "blue_horizon.data.augment_synthetic_data_with_gretel:main"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require external services (deselect with '-m \"not integration\"')",
]
